+++ CHANGELOG.md
@@ -35,6 +35,7 @@ Advantages:
 - [matterbridge]: Added brand to MatterbridgeAccessoryPlatform and MatterbridgeDynamicPlatform.
 - [PluginManager]: Verify brand of MatterbridgePlatform, MatterbridgeAccessoryPlatform and MatterbridgeDynamicPlatform.
 - [MatterbridgePlatform]: Verify brand of MatterbridgeEndpoint.
+- [BasicVideoPlayer]: Added Basic Video Player class.
 
 ### Changed